+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 03-07-2024
+### Added
+- Created a Form Generator Component
+    - added Validation Schema
+    - added API Hook Call
+    - added Callback Function for Form Rendering
+    - Both fields are displaying error messages if the input doesn't meet the
+    criteria set by the validation schema.
+    - used the POST method
+
 ## [1.0.2] - 02-07-2024
 ### Added
 - Created a List Component
@@ -19,8 +29,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - edited configuration
     - installed dependencies
     
-
-
 ## [1.0.0] - 02-07-2024
 ### Added
 - Initial Commit | Copy from homework-task